+++ cypress/support/e2e.js
@@ -18,3 +18,7 @@ import './commands'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
+
+// tab support
+import 'cypress-plugin-tab'
+