+++ src/custom-elements/src/constants/classes.ts
@@ -1,5 +1,6 @@
 //#region Common
 export const ACTIVE_CLASS = '-active';
+export const INACTIVE_CLASS = '-inactive';
 export const CLOSE_CLASS = '-close';
 export const CLOSED_CLASS = '-closed';
 export const LIGHT_CLASS = '-light';
@@ -25,7 +26,7 @@ export const ACCORDION_CLASSES = {
   CONTENT: 'chi-accordion__content',
   ITEM: 'chi-accordion__item',
   TRIGGER: 'chi-accordion__trigger',
-  TITLE: 'chi-accordion__title'
+  TITLE: 'chi-accordion__title',
 };
 //#endregion
 
@@ -39,14 +40,14 @@ export const BULK_ACTIONS_CLASSES = {
   BULK_ACTIONS_END: 'chi-bulk-actions__end',
   BULK_ACTIONS_BUTTONS: 'chi-bulk-actions__buttons',
   BULK_ACTIONS_BUTTONS_MOBILE: 'chi-bulk-actions__buttons-mobile',
-  BULK_ACTIONS_BUTTONS_DESKTOP: 'chi-bulk-actions__buttons-desktop'
+  BULK_ACTIONS_BUTTONS_DESKTOP: 'chi-bulk-actions__buttons-desktop',
 };
 //#endregion
 
 //#region Backdrop
 export const BACKDROP_CLASSES = {
   BACKDROP: 'chi-backdrop',
-  WRAPPER: 'chi-backdrop__wrapper'
+  WRAPPER: 'chi-backdrop__wrapper',
 };
 //#endregion
 
@@ -65,8 +66,8 @@ export const BUTTON_CLASSES = {
     SM: '-sm',
     MD: '-md',
     LG: '-lg',
-    XL: '-xl'
-  }
+    XL: '-xl',
+  },
 };
 //#endregion
 
@@ -87,7 +88,7 @@ export const CAROUSEL_CLASSES = {
   DOTS_ADDITION: '-dots',
   PAGINATION_ADDITION: '-pagination',
   DOT: 'chi-carousel__dot',
-  PAGINATION: 'chi-carousel__pagination'
+  PAGINATION: 'chi-carousel__pagination',
 };
 //#endregion
 
@@ -104,7 +105,10 @@ export const DROPDOWN_CLASSES = {
   TRIGGER: 'chi-dropdown__trigger',
   MENU: 'chi-dropdown__menu',
   MENU_CONTENT: 'chi-dropdown__menu-content',
-  MENU_ITEM: 'chi-dropdown__menu-item'
+  MENU_ITEM: 'chi-dropdown__menu-item',
+  MENU_ITEMS_WRAPPER: 'chi-dropdown__menu-items-wrapper',
+  MENU_HEADER: 'chi-dropdown__menu-header',
+  MENU_FOOTER: 'chi-dropdown__menu-footer',
 };
 //#endregion
 
@@ -113,14 +117,14 @@ export const CHECKBOX_CLASSES = {
   checkbox: 'chi-checkbox',
   INPUT: 'chi-checkbox__input',
   LABEL: 'chi-checkbox__label',
-  INDETERMINATE: '-indeterminate'
+  INDETERMINATE: '-indeterminate',
 };
 //#endregion
 
 //#region Divider
 export const DIVIDER_CLASSES = {
   DIVIDER: 'chi-divider',
-  VERTICAL: '-vertical'
+  VERTICAL: '-vertical',
 };
 //#endregion
 
@@ -128,7 +132,7 @@ export const DIVIDER_CLASSES = {
 export const RADIO_CLASSES = {
   RADIO: 'chi-radio',
   INPUT: 'chi-radio__input',
-  LABEL: 'chi-radio__label'
+  LABEL: 'chi-radio__label',
 };
 //#endregion
 
@@ -159,13 +163,13 @@ export const DATA_TABLE_CLASSES = {
   SORTING: '-sorting',
   COMPACT: '-compact',
   TRUNCATED: '-truncated',
-  PRINT_FOOTER: 'chi-data-table__print-footer'
+  PRINT_FOOTER: 'chi-data-table__print-footer',
 };
 //#endregion
 
 //#region Table
 export const TABLE_CLASSES = {
-  TABLE: 'chi-table'
+  TABLE: 'chi-table',
 };
 //#endregion
 
@@ -175,14 +179,14 @@ export const DRAWER_CLASSES = {
   HEADER: 'chi-drawer__header',
   TITLE: 'chi-drawer__title',
   CONTENT: 'chi-drawer__content',
-  TRIGGER: 'chi-drawer__trigger'
+  TRIGGER: 'chi-drawer__trigger',
 };
 //#endregion
 
 //#region Form
 export const FORM_CLASSES = {
   FORM_ITEM: 'chi-form__item',
-  LABEL: 'chi-label'
+  LABEL: 'chi-label',
 };
 //#endregion
 
@@ -192,7 +196,7 @@ export const MODAL_CLASSES = {
   MODAL: 'chi-modal',
   HEADER: 'chi-modal__header',
   CONTENT: 'chi-modal__content',
-  TITLE: 'chi-modal__title'
+  TITLE: 'chi-modal__title',
 };
 //#endregion
 
@@ -203,7 +207,7 @@ export const GENERIC_SIZE_CLASSES = {
   MD: '-md',
   LG: '-lg',
   XL: '-xl',
-  SM_2: '-sm--2'
+  SM_2: '-sm--2',
 };
 //#endregion
 
@@ -219,8 +223,8 @@ export const INPUT_CLASSES = {
   INPUT: 'chi-input',
   WRAPPER: 'chi-input__wrapper',
   ICON: {
-    RIGHT: '-icon--right'
-  }
+    RIGHT: '-icon--right',
+  },
 };
 //#endregion
 
@@ -235,25 +239,25 @@ export const PAGINATION_CLASSES = {
   START: 'chi-pagination__start',
   CENTER: 'chi-pagination__center',
   END: 'chi-pagination__end',
-  JUMPER: 'chi-pagination__jumper'
+  JUMPER: 'chi-pagination__jumper',
 };
 //#endregion
 
 //#region Search Input
 export const SEARCH_INPUT_CLASSES = {
-  SEARCH_INPUT: 'chi-search__input'
+  SEARCH_INPUT: 'chi-search__input',
 };
 //#endregion
 
 //#region Search Input
 export const SELECT_CLASSES = {
-  SELECT: 'chi-select'
+  SELECT: 'chi-select',
 };
 //#endregion
 
 //#region Tooltip
 export const TOOLTIP_CLASSES = {
-  TOOLTIP_ELEMENT: 'chi-tooltip'
+  TOOLTIP_ELEMENT: 'chi-tooltip',
 };
 //#endregion
 
@@ -268,14 +272,14 @@ export const TIME_CLASSES = {
   SECONDS: 'chi-time-picker__seconds',
   SECOND: 'chi-time-picker__second',
   PERIODS: 'chi-time-picker__periods',
-  PERIOD: 'chi-time-picker__period'
+  PERIOD: 'chi-time-picker__period',
 };
 //#endregion
 
 //#region Phone Input
 export const PHONE_INPUT_CLASSES = {
   PHONE_INPUT: 'chi-phone-input',
-  CODE: 'chi-phone-input__code'
+  CODE: 'chi-phone-input__code',
 };
 //#endregion
 
@@ -292,7 +296,7 @@ export const POPOVER_CLASSES = {
   NO_ARROW: '-no-arrow',
   POPOVER: 'chi-popover',
   TEXT: 'chi-popover__text',
-  TITLE: 'chi-popover__title'
+  TITLE: 'chi-popover__title',
 };
 //#endregion
 
@@ -307,7 +311,7 @@ export const TABS_CLASSES = {
   SOLID: '-solid',
   TRIGGER: 'chi-tabs__trigger',
   SHOW_MORE: 'chi-tabs__show-more',
-  VERTICAL: '-vertical'
+  VERTICAL: '-vertical',
 };
 //#endregion
 
@@ -316,23 +320,23 @@ export const UTILITY_CLASSES = {
   Z_INDEX: {
     Z_10: '-z--10',
     Z_30: '-z--30',
-    Z_40: '-z--40'
+    Z_40: '-z--40',
   },
   ALIGN_ITEMS: {
-    CENTER: '-align-items--center'
+    CENTER: '-align-items--center',
   },
   ALIGN_SELF: {
     CENTER: '-align-self--center',
-    LEFT: '-align-self--left'
+    LEFT: '-align-self--left',
   },
   DISPLAY: {
     FLEX: '-d--flex',
     SCREEN_ONLY: '-d-screen--only',
-    PRINT_ONLY: '-d-print--only'
+    PRINT_ONLY: '-d-print--only',
   },
   FLEX: {
     FLEX_GROW1: '-flex--grow1',
-    COLUMN: '-flex--column'
+    COLUMN: '-flex--column',
   },
   JUSTIFY: {
     CENTER: '-justify-content--center',
@@ -340,19 +344,19 @@ export const UTILITY_CLASSES = {
     END: '-justify-content--end',
     MD_START: '-justify-content-md--start',
     MD_CENTER: '-justify-content-md--center',
-    MD_END: '-justify-content-md--end'
+    MD_END: '-justify-content-md--end',
   },
   TYPOGRAPHY: {
     TEXT_TRUNCATE: '-text--truncate',
     TEXT_NORMAL: '-text--normal',
-    TEXT_BOLD: '-text--bold'
+    TEXT_BOLD: '-text--bold',
   },
   POSITION: {
     RELATIVE: '-position--relative',
-    ABSOLUTE: '-position--absolute'
+    ABSOLUTE: '-position--absolute',
   },
   SIZING: {
-    W100: '-w--100'
+    W100: '-w--100',
   },
   PADDING: {
     1: '-p--1',
@@ -364,13 +368,13 @@ export const UTILITY_CLASSES = {
       1: '-px--1',
       2: '-px--2',
       3: '-px--3',
-      4: '-px--4'
+      4: '-px--4',
     },
     Y: {
       1: '-py--1',
       2: '-py--2',
       3: '-py--3',
-      4: '-py--4'
+      4: '-py--4',
     },
     LEFT: {
       1: '-pl--1',
@@ -378,26 +382,26 @@ export const UTILITY_CLASSES = {
       3: '-pl--3',
       4: '-pl--4',
       5: '-pl--5',
-      6: '-pl--6'
+      6: '-pl--6',
     },
     RIGHT: {
       1: '-pr--1',
       2: '-pr--2',
       3: '-pr--3',
-      4: '-pr--4'
+      4: '-pr--4',
     },
     TOP: {
       1: '-pt--1',
       2: '-pt--2',
       3: '-pt--3',
-      4: '-pt--4'
+      4: '-pt--4',
     },
     BOTTOM: {
       1: '-pb--1',
       2: '-pb--2',
       3: '-pb--3',
-      4: '-pb--4'
-    }
+      4: '-pb--4',
+    },
   },
   MARGIN: {
     1: '-m--1',
@@ -408,40 +412,40 @@ export const UTILITY_CLASSES = {
       1: '-mx--1',
       2: '-mx--2',
       3: '-mx--3',
-      4: '-mx--4'
+      4: '-mx--4',
     },
     Y: {
       1: '-my--1',
       2: '-my--2',
       3: '-my--3',
-      4: '-my--4'
+      4: '-my--4',
     },
     LEFT: {
       1: '-ml--1',
       2: '-ml--2',
       3: '-ml--3',
-      4: '-ml--4'
+      4: '-ml--4',
     },
     RIGHT: {
       0: '-mr--0',
       1: '-mr--1',
       2: '-mr--2',
       3: '-mr--3',
-      4: '-mr--4'
+      4: '-mr--4',
     },
     TOP: {
       1: '-mt--1',
       2: '-mt--2',
       3: '-mt--3',
-      4: '-mt--4'
+      4: '-mt--4',
     },
     BOTTOM: {
       1: '-mb--1',
       2: '-mb--2',
       3: '-mb--3',
-      4: '-mb--4'
-    }
-  }
+      4: '-mb--4',
+    },
+  },
 };
 //#endregion
 
@@ -461,5 +465,19 @@ export const EPANEL_CLASSES = {
   TITLE: 'chi-epanel__title',
   STATE_ICON: '-state-icon',
   STATE: '-state',
+};
+//#endregion
+
+//#region Form
+export const DATEPICKER_CLASSES = {
+  DATEPICKER: 'chi-datepicker',
+  DAY: 'chi-datepicker__day',
+  DAY_NAMES: 'chi-datepicker__day-names',
+  DAYS: 'chi-datepicker__days',
+  WEEK_DAY: 'chi-datepicker__week-day',
+  MONTH: 'chi-datepicker__month',
+  MONTH_ROW: 'chi-datepicker__month-row',
+  PREV_MONTH: 'prev',
+  NEXT_MONTH: 'next',
 };
 //#endregion