+++ src/documentation/components/Navigation.vue
@@ -4,7 +4,7 @@
       <ThemeSwitcher />
     </div>
     <ul class="chi-tabs -vertical -lg -animated">
-      <li>
+      <li :class="{'-active' : $route.fullPath.includes(SECTION_URLS.gettingStarted)}">
         <div class="chi-collection__title">Getting Started</div>
         <ul class="chi-tabs__subtabs">
           <li>
@@ -30,7 +30,7 @@
           </li>
         </ul>
       </li>
-      <li>
+      <li :class="{'-active' : $route.fullPath.includes(SECTION_URLS.foundations)}">
         <div class="chi-collection__title">Foundations</div>
         <ul class="chi-tabs__subtabs">
           <li>
@@ -121,7 +121,7 @@
           </li>
         </ul>
       </li>
-      <li class="-active">
+      <li :class="{'-active' : $route.fullPath.includes(SECTION_URLS.components)}">
         <div class="chi-collection__title">Components</div>
         <ul class="chi-tabs__subtabs">
           <li
@@ -129,7 +129,7 @@
             v-for="(navigationComponentItem,
             index) in NAVIGATION_COMPONENTS_ITEMS"
             :class="{
-              '-active': $route.path === `/${navigationComponentItem.to}`
+              '-active': $route.path.includes(navigationComponentItem.to)
             }"
           >
             <NuxtLink
@@ -153,61 +153,30 @@
           </li>
         </ul>
       </li>
-      <li>
+      <li :class="{'-active' : $route.fullPath.includes(SECTION_URLS.utilities)}">
         <div class="chi-collection__title">Utilities</div>
         <ul class="chi-tabs__subtabs">
-          <li>
-            <a :href="`${BASE_URL}utilities/border`">Border</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/color`">Color</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/display`">Display</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/flex`">Flex</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/image`">Image</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/opacity`">Opacity</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/overflow`">Overflow</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/position`">Position</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/shadow`">Shadow</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/sizing`">Sizing</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/spacing`">Spacing</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/text`">Text</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/vertical-align`">Vertical-align</a>
-          </li>
-          <li>
-            <a :href="`${BASE_URL}utilities/zindex`">Z-index</a>
+          <li
+            :key="index"
+            v-for="(navigationUtilityItem, index) in NAVIGATION_UTILITIES_ITEMS"
+            :class="{
+              '-active': $route.path.includes(navigationUtilityItem.to)
+            }"
+          >
+            <NuxtLink :to="`/${navigationUtilityItem.to}`">
+              {{ navigationUtilityItem.label }}
+            </NuxtLink>
           </li>
         </ul>
       </li>
-      <li>
+      <li :class="{'-active' : $route.fullPath.includes(SECTION_URLS.templates)}">
         <div class="chi-collection__title">Templates</div>
         <ul class="chi-tabs__subtabs">
           <li
             :key="index"
             v-for="(navigationTemplateItem, index) in NAVIGATION_TEMPLATE_ITEMS"
             :class="{
-              '-active': $route.path === `/${navigationTemplateItem.to}`
+              '-active': $route.path.includes(navigationTemplateItem.to)
             }"
           >
             <NuxtLink
@@ -241,7 +210,9 @@ import ThemeSwitcher from './ThemeSwitcher.vue';
 import {
   NAVIGATION_COMPONENTS_ITEMS,
   NAVIGATION_TEMPLATE_ITEMS,
-  BASE_URL
+  NAVIGATION_UTILITIES_ITEMS,
+  BASE_URL,
+  SECTION_URLS
 } from '../constants/constants';
 
 declare const chi: any;
@@ -254,7 +225,9 @@ declare const chi: any;
     return {
       BASE_URL,
       NAVIGATION_COMPONENTS_ITEMS,
-      NAVIGATION_TEMPLATE_ITEMS
+      NAVIGATION_TEMPLATE_ITEMS,
+      NAVIGATION_UTILITIES_ITEMS,
+      SECTION_URLS
     };
   }
 })