+++ src/documentation/plugins/chi-docs.js
@@ -1,14 +1,13 @@
 import Vue from 'vue';
 import axios from 'axios';
-import { DOCS_URL } from '../constants/constants';
 
 async function getDocs() {
   return await axios
-    .get(DOCS_URL)
-    .then(response => {
+    .get(`${process.env.BASE_URL}js/ce/docs.json`)
+    .then((response) => {
       return response.data;
     })
-    .catch(error => console.log(error));
+    .catch((error) => console.log(error));
 }
 
 export default async () => {